test: fix several tests

+++ packages/datasource-replica/src/synchronization/cache-target.ts
@@ -6,7 +6,7 @@ import type {
   ResolvedOptions,
   SynchronizationTarget,
 } from '../types';
-import type { RecordData } from '@forestadmin/datasource-toolkit';
+import type { Logger, RecordData } from '@forestadmin/datasource-toolkit';
 import type { Sequelize } from 'sequelize';
 
 import { Op } from 'sequelize';
@@ -17,10 +17,12 @@ import { escape, getRecordId } from '../utils';
 export default class CacheTarget implements SynchronizationTarget {
   connection: Sequelize;
   options: ResolvedOptions;
+  logger: Logger;
 
-  constructor(connection: Sequelize, options: ResolvedOptions) {
+  constructor(connection: Sequelize, options: ResolvedOptions, logger: Logger) {
     this.connection = connection;
     this.options = options;
+    this.logger = logger;
   }
 
   async applyDump(changes: PullDumpResponse, firstPage: boolean): Promise<void> {
@@ -38,13 +40,13 @@ export default class CacheTarget implements SynchronizationTarget {
     const entries = changes.entries.flatMap(entry => this.flattenRecord(entry));
 
     for (const entry of entries) {
-      if (!recordsByCollection[entry.collection]) recordsByCollection[entry.collection] = [];
-      recordsByCollection[entry.collection].push(entry.record);
+      if (entry !== undefined) {
+        if (!recordsByCollection[entry.collection]) recordsByCollection[entry.collection] = [];
+        recordsByCollection[entry.collection].push(entry.record);
+      }
     }
 
     for (const [collection, records] of Object.entries(recordsByCollection)) {
-      this.checkCollection(collection);
-
       // fixme check that no pre-processing of the records are needed to handle
       // dates, buffers, ...
       await this.connection.model(collection).bulkCreate(records);
@@ -57,20 +59,27 @@ export default class CacheTarget implements SynchronizationTarget {
     // [transaction]
 
     for (const entry of changes.deletedEntries) {
-      this.checkCollection(entry.collection);
-      await this.destroySubModels(entry);
-      await this.connection.model(entry.collection).destroy({ where: entry.record });
+      if (
+        this.checkCollectionAndFieldsInSchema(entry.collection, Object.keys(entry.record), false)
+      ) {
+        await this.destroySubModels(entry);
+        await this.connection.model(entry.collection).destroy({ where: entry.record });
+      }
     }
 
     // Upsert records in both root and virtual collections
-    for (const entry of changes.newOrUpdatedEntries) {
-      this.checkCollection(entry.collection);
-      await this.destroySubModels(entry);
 
-      // fixme check that no pre-processing of the records are needed to handle
-      // dates, buffers, ...
-      for (const subEntry of this.flattenRecord(entry))
-        await this.connection.model(subEntry.collection).upsert(subEntry.record);
+    for (const entry of changes.newOrUpdatedEntries) {
+      if (
+        this.checkCollectionAndFieldsInSchema(entry.collection, Object.keys(entry.record), false)
+      ) {
+        await this.destroySubModels(entry);
+
+        // fixme check that no pre-processing of the records are needed to handle
+        // dates, buffers, ...
+        for (const subEntry of this.flattenRecord(entry))
+          await this.connection.model(subEntry.collection).upsert(subEntry.record);
+      }
     }
 
     // [/transaction]
@@ -79,12 +88,19 @@ export default class CacheTarget implements SynchronizationTarget {
   private flattenRecord(
     recordWithCollection: RecordDataWithCollection,
   ): RecordDataWithCollection[] {
-    return flattenRecord(
-      recordWithCollection,
-      this.options.schema.find(s => s.name === recordWithCollection.collection).fields,
-      this.options?.flattenOptions?.[recordWithCollection.collection]?.asFields ?? [],
-      this.options?.flattenOptions?.[recordWithCollection.collection]?.asModels ?? [],
-    );
+    if (
+      this.checkCollectionAndFieldsInSchema(
+        recordWithCollection.collection,
+        Object.keys(recordWithCollection.record),
+        false,
+      )
+    )
+      return flattenRecord(
+        recordWithCollection,
+        this.options.schema.find(s => s.name === recordWithCollection.collection).fields,
+        this.options?.flattenOptions?.[recordWithCollection.collection]?.asFields ?? [],
+        this.options?.flattenOptions?.[recordWithCollection.collection]?.asModels ?? [],
+      );
   }
 
   private async destroySubModels(entry: RecordDataWithCollection): Promise<void> {
@@ -101,8 +117,58 @@ export default class CacheTarget implements SynchronizationTarget {
     await Promise.all(promises);
   }
 
-  private checkCollection(collection: string): void {
+  private checkCollection(collection: string, shouldTrow: boolean): void {
     if (!this.options.schema.find(s => s.name === collection))
-      throw new Error(`Collection ${collection} not found in schema`);
+      if (shouldTrow) {
+        throw new Error(`Collection ${collection} not found in schema`);
+      } else {
+        this.logger('Error', `Collection ${collection} not found in schema`);
+      }
+  }
+
+  private checkCollectionAndFieldsInSchema(
+    collection: string,
+    fields: string[],
+    shouldThrow: boolean,
+  ): boolean {
+    const schemaEntry = this.options.schema.find(s => s.name === collection);
+
+    if (!schemaEntry) {
+      if (shouldThrow) {
+        throw new Error(`Collection ${collection} not found in schema`);
+      } else {
+        this.logger('Error', `Collection '${collection}' not found in schema`);
+      }
+
+      return false; // stop the flatten for this entry
+    }
+
+    const schemaFields = Object.keys(schemaEntry.fields);
+
+    const missingFields = fields.filter(field => !schemaFields.includes(field));
+
+    if (fields.length - missingFields.length <= 0) {
+      this.logger(
+        'Error',
+        // eslint-disable-next-line max-len
+        `Cannot find any field in the given records matching the schema of collection "${collection}"`,
+      );
+
+      return false;
+    }
+
+    if (missingFields.length > 0) {
+      const errorMessage = `Fields '${missingFields.join(
+        ', ',
+      )}' do not exist in the schema for collection ${collection}`;
+
+      if (shouldThrow) {
+        throw new Error(errorMessage);
+      } else {
+        this.logger('Warn', errorMessage);
+      }
+    }
+
+    return true; // continue flatten for this entry
   }
 }
